Added check for readonly and initialization to CC0031 + some unit tests.
